Moved Add and Refresh button to Toolbar, combined standingCtrl and adminCtrl for mainCtrl
